requirements.txt
# Core dependencies
torch>=1.12.0
torchaudio>=0.12.0
transformers>=4.20.0
librosa>=0.9.0
soundfile>=0.10.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Web framework
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
pathlib2>=2.3.0

# Optional dependencies for enhanced functionality
# jiwer>=2.3.0  # For better WER calculation
# matplotlib>=3.5.0  # For plotting
# seaborn>=0.11.0  # For visualization
# tensorboard>=2.8.0  # For training visualization